requirements.txt
pytesseract>=0.3.10
opencv-python>=4.10.0.84
Pillow>=10.4.0
numpy>=2.1.0

